reddit_radar_api_server/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
praw>=7.7.0
python-dateutil>=2.8.2
pydantic>=2.5.0
pandas>=2.0.0
